requirements.txt
groq
python-dotenv
discord.py